+++ etc/install.sh
@@ -2,72 +2,94 @@
 # coding: utf-8
 # (c) 2016 Dalton Serey, UFCG
 #
-# Interactive TST Installer script. Download and install tst CLI scripts. 
+# Interactive TST Installer script. Run this script to download
+# and install TST CLI tools. This script can be invoked with
+# these options:
+#
+# --pre-release
+#       Download the latest pre-release version available.
+#
+# --update
+#       Update existing installation. In this mode, the install
+#       script runs in non-interactive mode. It prints less
+#       evolution messages, it doesn't ask whether or not to
+#       overwrite previous installations, it doesn't configure
+#       the enviroment and doesn't delete old installations of
+#       tst.  This mode is used by the tst update command.
+#
+# --installation-dir <dir>
+#       Install the new version into <dir>.
 
 # constants
 INSTALL_DIR=~/.tst.install
 TST_DIR=~/.tst
 CONFIG_FILE=~/.tst/config.json
-INTERACTIVE="true"
 
 # colors
-RESET="\033[0m"
-BLACK="\033[0;30m"
-BLUE="\033[0;34m"
-BROWN="\033[0;33m"
-CYAN="\033[0;36m"
-DGRAY="\033[1;30m"
-GREEN="\033[0;32m"
-LBLUE="\033[1;34m"
-LCYAN="\033[1;36m"
 LGRAY="\033[0;37m"
-LGREEN="\033[1;32m"
-LPURPLE="\033[1;35m"
 LRED="\033[1;31m"
-PURPLE="\033[0;35m"
-RED="\033[0;31m"
-WHITE="\033[1;37m"
-YELLOW="\033[1;33m"
+LBLUE="\033[1;34m"
+LGREEN="\033[1;32m"
 
 # semantic colors
+RESET="\033[0m"
 NORMAL=$LGRAY
 WARNING=$LRED
 IMPORTANT=$LBLUE
 QUESTION=$LGREEN
 
-# read either 'y' or 'n' from keyboard
-function get_yes_or_no {
-    while true; do 
-        read -s -n 1 ANSWER
-        if [ "$ANSWER" == 'y' ]; then break; fi
-        if [ "$ANSWER" == 'n' ]; then break; fi
-    done
-    echo $ANSWER
-}
-
 # print with color
 function print {
     COLOR=$2
-    if [ "$COLOR" == "" ]; then
+    if [[ "$COLOR" == "" ]]; then
         COLOR=$NORMAL
     fi
 
     echo -n -e $COLOR"$1"$RESET
 }
 
+# read either 'y' or 'n' from keyboard
+function get_yes_or_no {
+    while true; do 
+        read -s -n 1 answer
+        [[ "$answer" == "y" ]] && break
+        [[ "$answer" == "n" ]] && break
+    done
+    echo $answer
+}
+
+# locate command or abort
+function require_command {
+    local command_name=$1
+    locate_command=$(command -v $command_name)
+    if [ $? != 0 ]; then
+        print "The installation script requires the $command_name command\n" $WARNING
+        print "Aborting installation\n"
+        exit 1
+    fi
+}
 
 # MAIN
+if [[ "$EUID" == "0" ]]; then
+   print "This script cannot be run as root\n" $WARNING
+   exit 1
+fi
+
+# require curl and unzip
+require_command curl
+require_command unzip
+
 # process options
+mode="installation"
+verbose="true"
 while (( $# > 0 )); do
     case "$1" in
-        --del-previous)
-            DELETE_PREVIOUS="true"
+        --pre-release)
+            GET_PRE_RELEASE="true"
             ;;
-        --development-version)
-            DOWNLOAD_DEV_VERSION="true"
-            ;;
-        --non-interactive)
-            INTERACTIVE="false"
+        --update)
+            mode="update"
+            verbose="false"
             ;;
         --installation-dir)
             INSTALLATION_DIR="true"
@@ -82,110 +104,73 @@ while (( $# > 0 )); do
     shift
 done
 
-# require curl or abort
-CURL=$(command -v curl)
-if [ $? != 0 ]; then
-    echo "the installation script requires the curl command";
-    echo "aborting installation";
-    exit 1
-fi
-
-# require unzip or abort
-UNZIP=$(command -v unzip)
-if [ $? != 0 ]; then
-    echo "the installation script requires the unzip command";
-    echo "aborting installation";
-    exit 1
-fi
-
-# identify releases url
-if [ "$DOWNLOAD_DEV_VERSION" == "true" ]; then
-    RELEASES_URL='https://api.github.com/repos/daltonserey/tst/releases'
-    if [ "$INTERACTIVE" == "true" ]; then
-        print "* fetching development pre-release information\n"
-    fi
+# set releases url
+print "Starting tst $mode\n" $IMPORTANT
+if [ "$GET_PRE_RELEASE" == "true" ]; then
+    releases_url='https://api.github.com/repos/daltonserey/tst/releases'
+    [[ "$verbose" == "true" ]] && print "* cheking available releases\n"
 else
-    RELEASES_URL='https://api.github.com/repos/daltonserey/tst/releases/latest'
-    if [ "$INTERACTIVE" == "true" ]; then
-        print "* fetching latest release information\n"
-    fi
+    releases_url='https://api.github.com/repos/daltonserey/tst/releases/latest'
+    [[ "$verbose" == "true" ]] && print "* fetching latest release information\n"
 fi
 
-# download releases info: identify tag_name and zipball_url
-RELEASES=$(curl -q $RELEASES_URL 2> /dev/null)
-if [ $? != 0 ]; then
+# download releases info and parse tag_name and zipball_url
+releases=$(curl -q $releases_url 2> /dev/null)
+if [[ $? != 0 ]]; then
     print "Couldn't download release information\n" $WARNING
-    print "Installation aborted\n"
+    print "$mode aborted\n"
     exit 1
 fi
-TAG_NAME=$(echo -e "$RELEASES" | grep "tag_name" | cut -f 4 -d '"' | head -1)
-ZIPBALL_URL=$(echo -e "$RELEASES" | grep "zipball_url" | cut -f 4 -d '"' | head -1)
+tag_name=$(echo -e "$releases" | grep "tag_name" | cut -f 4 -d '"' | head -1)
+zipball_url=$(echo -e "$releases" | grep "zipball_url" | cut -f 4 -d '"' | head -1)
 
 # cancel installation if there's no release available
-if [ "$TAG_NAME" == "" ]; then
+if [[ "$tag_name" == "" ]]; then
     print "No release available\n" $WARNING
-    print "Installation canceled\n" $IMPORTANT
+    print "$mode canceled\n" $IMPORTANT
     exit 1
+else
+    print "Latest available release: $tag_name\n" $IMPORTANT
 fi
 
-# create TST_DIR if it doesn't exist
-mkdir -p $TST_DIR
-
-# check for previous installation version
-if [ -f "$TST_DIR/release.json" ]; then
-    PREVIOUS_TAG_NAME=$(cat $TST_DIR/release.json | grep "tag_name" | cut -f 4 -d '"')
-
-    # notify user about previous installation
-    if [ "$PREVIOUS_TAG_NAME" == "$TAG_NAME" ]; then
-        print "Installed tst is up-to-date (version $TAG_NAME)\n" $IMPORTANT
-        exit
-    else
-        print "New version of tst available (version $TAG_NAME)\n" $IMPORTANT
-    fi
-
-    if [ "$INTERACTIVE" == "true" ]; then
-        # ask user whether to proceed and overwrite installation
-        print "Proceed and overwrite? (y/n) " $QUESTION
-        get_yes_or_no
-        if [ "$ANSWER" != "y" ]; then
-            print "Installation cancelled by user\n" $IMPORTANT
-            exit 0
-        fi
+# if in installation mode, check for previous installation
+if [[ "$mode" == "installation" ]] && [[ -d "$TST_DIR" ]]; then
+    print "\nAn installation of tst was found\n" $WARNING
+    print "Overwrite? (y/n) " $QUESTION
+    get_yes_or_no
+    if [ "$answer" == "n" ]; then
+        print "$mode cancelled by user\n"
+        exit 0
     fi
 fi
 
-# create new installation dir
-if [ -d "$INSTALL_DIR" ]; then
-    if [ "$INTERACTIVE" == "true" ]; then
-        print "* deleting failed attempt to install" $WARNING 
-    fi
+# create and cd to new installation dir
+if [[ -d "$INSTALL_DIR" ]]; then
+    [[ "$verbose" == "true" ]] && print "* deleting files from previous installation attempt\n" $WARNING
     rm -rf $INSTALL_DIR
 fi
 mkdir -p $INSTALL_DIR
+cd $INSTALL_DIR
 
 # download latest release into INSTALL_DIR
-cd $INSTALL_DIR
-if [ "$INTERACTIVE" == "true" ]; then
-    print "* downloading release zip\n"
-fi
-curl -q -Lko tst.zip $ZIPBALL_URL 2> /dev/null
-if [ $? != 0 ]; then
+[[ "$verbose" == "true" ]] && print "* downloading $tag_name zip\n"
+curl -q -Lko tst.zip $zipball_url &> $INSTALL_DIR/log
+if [[ $? != 0 ]]; then
     rm -rf $INSTALL_DIR
-    echo $ZIPBALL_URL
+    echo $zipball_url
     print "Couldn't download release zip\n" $WARNING
-    print "Installation aborted\n"
     print "Temporary files deleted\n"
+    print "$mode aborted\n"
     exit 1
 fi
 
-# unzip and install tst scripts within INSTALL_DIR
-if [ "$INTERACTIVE" == "true" ]; then
-    print "* unzipping and installing tst scripts\n"
-fi
+# unzip tst scripts
+[[ "$verbose" == "true" ]] && print "* unzipping installation files\n"
 unzip -q tst.zip
-rm tst.zip
 
-# install files in TST_DIR
+# install tst
+[[ "$verbose" == "true" ]] && print "* installing tst scripts\n"
+mkdir -p $TST_DIR
 mkdir -p $TST_DIR/bin
 mv daltonserey-tst*/bin/* $TST_DIR/bin/
 mkdir -p $TST_DIR/commands
@@ -198,21 +183,37 @@ mv daltonserey-tst*/LICENSE $TST_DIR/
 
 # update release.json in TST_DIR
 cd $TST_DIR
-echo "{\"tag_name\": \"$TAG_NAME\"}" > $TST_DIR/release.json
+echo "{\"tag_name\": \"$tag_name\"}" > $TST_DIR/release.json
 
-# end installation
+# finish installation/update
 rm -rf $INSTALL_DIR
-print "Installation finished\n" $IMPORTANT
+print "Finished $mode\n" $IMPORTANT
 
-# configure environment if in interactive mode
-if [ "$INTERACTIVE" == "true" ]; then
-    print "\nConfigure environment? (y/n) " $QUESTION
+# end script if this is an update
+[[ "$mode" == "update" ]] && exit 0
+
+
+# The remaining of this script is for full installations
+
+# delete/rename previous installation
+OLD_TST=~/tst
+if [[ -d "$OLD_TST" ]]; then
+    print "\nWe found what seems to be an older installation.\n" $IMPORTANT
+    print "$OLD_TST\n" $WARNING
+    print "Delete? (y/n) " $QUESTION
     get_yes_or_no
-    if [ "$ANSWER" == "y" ]; then
-        $TST_DIR/etc/setenv.sh
-    else
-        print "Environment was$WARNING not $NORMAL configured.\n"
-        print "Remember to add $IMPORTANT$TST_DIR/bin $NORMAL to your PATH\n"
-        exit
-    fi
+    [[ "$answer" == "y" ]] && rm -rf $OLD_TST
+fi
+
+# configure environment
+print "\nConfigure environment? (y/n) " $QUESTION
+get_yes_or_no
+if [[ "$answer" == "y" ]]; then
+    $TST_DIR/etc/setenv.sh
+    print "Finished environment configuration\n" $IMPORTANT
+    print "\nTo make configuration take effect immediately, type the command:\n"
+    print "*$IMPORTANT source ~/.bashrc$NORMAL\n"
+else
+    print "Environment was$WARNING not$NORMAL configured.\n"
+    print "Remember to add $IMPORTANT$TST_DIR/bin$NORMAL to PATH and PYTHONPATH\n"
 fi